Make tests pass for mysql adapter
Walked through the source to let the tests pass for the mysql adapter,
and partially the Redis adapter. This is for issue #5.

For this to work, some adapter methods were added to impelement more of CRUD,
as requested in #2 and #3.

Still to do: update and delete methods for the Redis adapter.